+++ src/refactoring/hooks/useCouponForm.tsx
@@ -1,0 +1,27 @@
+import { useState } from 'react';
+import { Coupon } from '../../types';
+
+export const useCouponForm = (onCouponAdd: (newCoupon: Coupon) => void) => {
+  const [newCoupon, setNewCoupon] = useState<Coupon>({
+    name: '',
+    code: '',
+    discountType: 'percentage',
+    discountValue: 0,
+  });
+
+  const handleAddCoupon = () => {
+    onCouponAdd(newCoupon);
+    setNewCoupon({
+      name: '',
+      code: '',
+      discountType: 'percentage',
+      discountValue: 0,
+    });
+  };
+
+  return {
+    newCoupon,
+    setNewCoupon,
+    handleAddCoupon,
+  };
+};
